(fix) resolve the dependency cycle error
